+++ src/main/java/com/connor/hozon/bom/resources/service/bom/impl/HzPbomServiceImpl.java
@@ -16,6 +16,7 @@ import com.connor.hozon.bom.resources.mybatis.bom.HzMbomRecordDAO;
 import com.connor.hozon.bom.resources.mybatis.bom.HzPbomRecordDAO;
 import com.connor.hozon.bom.resources.mybatis.bom.impl.HzPbomRecordDAOImpl;
 import com.connor.hozon.bom.resources.page.Page;
+import com.connor.hozon.bom.resources.query.HzBomRecycleByPageQuery;
 import com.connor.hozon.bom.resources.query.HzPbomByPageQuery;
 import com.connor.hozon.bom.resources.query.HzPbomTreeQuery;
 import com.connor.hozon.bom.resources.service.bom.HzPbomService;
@@ -245,7 +246,9 @@ public class HzPbomServiceImpl implements HzPbomService {
                     deleteHzPbomReqDTO.seteBomPuid(s);
                     list.add(deleteHzPbomReqDTO);
                 }
-                hzPbomRecordDAO.deleteList(list);//mabatis 做批量更新时 返回值为-1 所以这里根据异常情况来判断成功与否
+                if(ListUtil.isNotEmpty(list)){
+                    hzPbomRecordDAO.deleteList(list);//mabatis 做批量更新时 返回值为-1 所以这里根据异常情况来判断成功与否
+                }
             }
             return OperateResultMessageRespDTO.getSuccessResult();
         }catch (Exception e){
@@ -388,9 +391,9 @@ public class HzPbomServiceImpl implements HzPbomService {
                 return null;
             }
             List<HzPbomLineRecord> records = recordPage.getResult();
-            int num = (query.getPage()-1)*query.getLimit();
+            int num = (recordPage.getPageNumber()-1)*recordPage.getPageSize();
             List<HzPbomLineRespDTO> respDTOS = pbomLineRecordToRespDTOS(records,query.getProjectId(),num);
-            return new Page<>(query.getPage(),query.getLimit(),recordPage.getTotalCount(),respDTOS);
+            return new Page<>(recordPage.getPageNumber(),recordPage.getPageSize(),recordPage.getTotalCount(),respDTOS);
         }catch (Exception e){
             return null;
         }
@@ -712,6 +715,82 @@ public class HzPbomServiceImpl implements HzPbomService {
         }
     }
 
+    @Override
+    public Page<HzPbomLineRespDTO> getHzPbomRecycleByPage(HzBomRecycleByPageQuery query) {
+        try {
+            Page<HzPbomLineRecord> recordPage =hzPbomRecordDAO.getHzPbomRecycleRecord(query);
+            if(recordPage == null || recordPage.getResult() == null){
+                return null;
+            }
+            List<HzPbomLineRecord> records = recordPage.getResult();
+            int num = (recordPage.getPageNumber()-1)*recordPage.getPageSize();
+            List<HzPbomLineRespDTO> respDTOS = pbomLineRecordToRespDTOS(records,query.getProjectId(),num);
+            return new Page<>(recordPage.getPageNumber(),recordPage.getPageSize(),recordPage.getTotalCount(),respDTOS);
+        }catch (Exception e){
+            return null;
+        }
+    }
+
+    @Override
+    public OperateResultMessageRespDTO RecoverDeletePbomRecord(String projectId, String puid) {
+        OperateResultMessageRespDTO respDTO = new OperateResultMessageRespDTO();
+        try{
+            if(projectId == null || projectId == "" || puid == null || puid == ""){
+                respDTO.setErrCode(OperateResultMessageRespDTO.FAILED_CODE);
+                respDTO.setErrMsg("非法参数");
+                return  respDTO;
+            }
+            Map<String,Object> map = new HashMap<>();
+            map.put("projectId",projectId);
+            map.put("pPuid",puid);
+            List<HzPbomLineRecord> recordList = hzPbomRecordDAO.getPbomById(map);
+            if(ListUtil.isNotEmpty(recordList)){
+                respDTO.setErrMsg("当前要恢复对象已存在bom系统中！");
+                respDTO.setErrCode(OperateResultMessageRespDTO.FAILED_CODE);
+                return  respDTO;
+            }
+            map.put("status",0);//已删除的bom
+            recordList =hzPbomRecordDAO.getPbomById(map);
+            if(ListUtil.isNotEmpty(recordList)){
+                HzPbomLineRecord record = recordList.get(0);
+                if(record.getLineIndex().split("\\.").length == 2){
+                    respDTO.setErrMsg("2Y层结构无法恢复！");
+                    respDTO.setErrCode(OperateResultMessageRespDTO.FAILED_CODE);
+                    return  respDTO;
+                }
+                Map<String,Object> map1 = new HashMap<>();
+                map1.put("projectId",projectId);
+                map1.put("pPuid",record.getParentUid());
+                recordList = hzPbomRecordDAO.getPbomById(map1);
+                if(ListUtil.isEmpty(recordList)){
+                    respDTO.setErrMsg("当前要恢复对象的父结构不存在，无法恢复！");
+                    respDTO.setErrCode(OperateResultMessageRespDTO.FAILED_CODE);
+                    return  respDTO;
+                }else{
+                    if(recordList.get(0).getIsHas().equals(0)){
+                        HzPbomLineRecord lineRecord = recordList.get(0);
+                        lineRecord.setIsHas(1);
+                        lineRecord.setIsPart(0);
+                        if(lineRecord.getLineIndex().split("\\.").length == 2 && lineRecord.getIs2Y().equals(0)){
+                            lineRecord.setIs2Y(1);
+                        }
+                        int i = hzPbomRecordDAO.update(lineRecord);
+                        if(i<=0){
+                            return OperateResultMessageRespDTO.getFailResult();
+                        }
+                    }
+                    int i = hzPbomRecordDAO.recoverBomById(record.geteBomPuid());
+                    if(i>0){
+                        return OperateResultMessageRespDTO.getSuccessResult();
+                    }
+                }
+            }
+            return OperateResultMessageRespDTO.getFailResult();
+        }catch (Exception e){
+            return OperateResultMessageRespDTO.getFailResult();
+        }
+    }
+
     private HzPbomLineRecord bomLineToPbomLine(HzBomLineRecord record){
         HzPbomLineRecord hzPbomLineRecord = new HzPbomLineRecord();
         hzPbomLineRecord.setPuid(UUID.randomUUID().toString());
@@ -825,6 +904,8 @@ public class HzPbomServiceImpl implements HzPbomService {
                 respDTO.setOuterPart(record.getOuterPart());
                 respDTO.setStation(record.getStation());
                 respDTO.setOrderNum(record.getOrderNum());
+                respDTO.setpBomLinePartName(record.getpBomLinePartName());
+                respDTO.setpBomLinePartEnName(record.getpBomLinePartEnName());
                 respDTOS.add(respDTO);
             }
             return respDTOS;
@@ -834,4 +915,5 @@ public class HzPbomServiceImpl implements HzPbomService {
         }
         return null;
     }
+
 }
